requirements.txt
# LangChain y LangGraph
langchain==0.3.13
langchain-openai==0.2.14
langgraph==0.2.60
langchain-core==0.3.28

# Base de datos SQL Server
pyodbc==5.2.0
sqlalchemy==2.0.36
pandas==2.2.3

# Utilidades
python-dotenv==1.0.1

# Opcional: para mejor visualización
colorama==0.4.6